Refactor casting (#30)
Refactors casting a bit. Fixes a bug where arrays were not being cast
correctly. I think this won't really scale to more complex types and we
should consider a dependency to handle this or code up a more
runtime dependent solution. But for now i'm aiming to
get it working well for the simple cases. Still have some tests to do.